+++ client/src/components/productReviews/productReviews.module.css
@@ -67,11 +67,17 @@
 
 .review-modal {
   width: 50%;
-  height: 50%;
+  height: 70%;
   border: 1px solid #dbe7f7ff;
   padding: 10px;
   margin-left: auto;
   margin-right: auto;
+  position: fixed;
+  z-index: 100;
+  display: block;
+  top: 10%;
+  left: 23%;
+  background-color: white;
 }
 
 .review-modal-content {
