+++ client/src/components/productReviews/test/productBreakdown.test.jsx
@@ -1,21 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer from '../../../reducers/rootReducer';
-import testReview from '../../../fixtures/testReview.json';
+import testStore from '../../../fixtures/testStore';
 import ProductBreakdown from '../productBreakdown.jsx';
 
 describe('productBreakdown', () => {
-  const testStore = createStore(
-    rootReducer,
-    {
-      reviews: testReview.reviews,
-      reviewMeta: testReview.reviewMeta,
-    },
-    applyMiddleware(thunk),
-  );
   it('should render a list of characteristics', () => {
     const { getByText, queryByText } = render(
       <Provider store={testStore}>
