+++ client/src/components/productReviews/test/productReviews.test.jsx
@@ -2,24 +2,14 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
 import fetchMock from 'jest-fetch-mock';
-import rootReducer from '../../../reducers/rootReducer';
 import testReview from '../../../fixtures/testReview.json';
+import testStore from '../../../fixtures/testStore';
 
 import ProductReviews from '../productReviews.jsx';
 import '../../common/fontAwesomeIcons';
 
 describe('productReviews', () => {
-  const testStore = createStore(
-    rootReducer,
-    {
-      reviews: testReview.reviews,
-      reviewMeta: testReview.reviewMeta,
-    },
-    applyMiddleware(thunk),
-  );
   beforeEach(() => {
     fetchMock.mockIf('http://127.0.0.1:3000', (req) => {
       if (req.url.endsWith('/reviews')) {
