+++ client/src/components/productReviews/test/ratings.test.jsx
@@ -2,21 +2,11 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer from '../../../reducers/rootReducer';
-import testReview from '../../../fixtures/testReview.json';
+import testStore from '../../../fixtures/testStore';
 import Ratings from '../ratings.jsx';
+import '../../common/fontAwesomeIcons';
 
 describe('Ratings', () => {
-  const testStore = createStore(
-    rootReducer,
-    {
-      reviews: testReview.reviews,
-      reviewMeta: testReview.reviewMeta,
-    },
-    applyMiddleware(thunk),
-  );
   it('should render the rating section without crash', () => {
     const { getByText } = render(
       <Provider store={testStore}>
