+++ client/src/components/productReviews/test/starBreakdown.test.jsx
@@ -1,21 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer from '../../../reducers/rootReducer';
-import testReview from '../../../fixtures/testReview.json';
 import StarBreakdown from '../starBreakdown.jsx';
+import testStore from '../../../fixtures/testStore';
 
 describe('starBreakdown', () => {
-  const testStore = createStore(
-    rootReducer,
-    {
-      reviews: testReview.reviews,
-      reviewMeta: testReview.reviewMeta,
-    },
-    applyMiddleware(thunk),
-  );
   it('should render the distribution of star reviews', () => {
     const { getByText } = render(
       <Provider store={testStore}>
